внес правки нейминга и почистил код
