+++ index.html
@@ -115,7 +115,7 @@
 
                 <!-- Добавление хэш-тегов и комментария к изображению -->
                 <fieldset class="img-upload__text text">
-                  <input class="text__hashtags" name="hashtags" placeholder="#ХэшТег" value="#ХэшТег" minlength="2">
+                  <input class="text__hashtags" name="hashtags" placeholder="#ХэшТег" >
                   <textarea class="text__description" name="description" placeholder="Ваш комментарий..." ></textarea>
                 </fieldset>
